update auth routs, update auth logic

+++ routes/users.js
@@ -1,13 +1,13 @@
 const express = require('express');
 const router = express.Router();
 const controller = require('../controllers/users');
-const encrypt = require('../utils/encrypt');
+const authenticateToken = require('../middlewares/authenticateToken');
 
 /* GET users listing. */
 router.get('/users', async (req, res, next) => {
   const users = await controller.getUsers();
 
-  res.json(users);
+  res.status(201).json(users);
 });
 
 /* GET user. */
@@ -15,46 +15,59 @@ router.get('/users/:id', async (req, res, next) => {
   const id = req.params.id;
   const user = await controller.getUserById(id);
 
-  res.json(user);
+  res.status(201).json(user);
 });
 
 /* POST user create. */
-router.post('/users', async (req, res, next) => {
-  const { name, email, password } = req.body;
-  const encryptedPassword = await encrypt.cryptPassword(password);
-  const createdUsers = await controller.createUser({ name, email, password: encryptedPassword });
+// Remove after auth will be finished
+// router.post('/users', async (req, res, next) => {
+//   const { name, email, password } = req.body;
+//   const encryptedPassword = await encrypt.cryptPassword(password);
+//   const createdUsers = await controller.createUser({ name, email, password: encryptedPassword });
 
-  res.json(createdUsers);
-});
+//   res.json(createdUsers);
+// });
 
-/* PATCH user. */
-router.put('/users', async (req, res, next) => {
-  const { name, email, id } = req.body;
+/* PUT user. */
+router.put('/users/:id', authenticateToken, async (req, res, next) => {
+  const { name, email } = req.body;
+  const id = req.params.id;
   const updatedUser = await controller.updateUser({ name, email, id });
 
-  res.json(updatedUser);
+  res.status(201).json(updatedUser);
 });
 
-// TODO: add patch for user,
 // TODO: change (patching) password
 /* PATCH user. */
-router.patch('/users/:id', function(req, res, next) {
-  res.send('respond with a resource');
+router.patch('/users/:id', authenticateToken, async (req, res, next) => {
+  const { name, email } = req.body;
+  const id = req.params.id;
+  const updatedUser = await controller.updateUser2({ name, email, id });
+
+  if (updatedUser) {
+    res.status(201).json(updatedUser);
+  } else {
+    res.status(500).send('Internal server error');
+  }
 });
 
 /* DELETE user. */
-router.delete('/users/:id', async (req, res, next) => {
+router.delete('/users/:id', authenticateToken, async (req, res, next) => {
   const id = req.params.id;
-  const result = await controller.deleteUser(id);
+  if (req.userId === id) {
+    const result = await controller.deleteUser(id);
 
-  if (result.rowCount > 0) {
-    res.send('User successfully deleted!');
+    if (result.rowCount > 0) {
+      res.status(201).send('User successfully deleted!');
+    } else {
+      res.status(500).send('Something went wrong!');
+    }
+  
   } else {
-    res.send('There is no user with this id!');
+    res.status(500).send('You don\'t have permissions to delete this user!');
   }
+  
 });
 // TODO: add error for bad request
-// TODO: auth (login, logout, signup) for users
-// TODO: jwt token logic for auth
 
 module.exports = router;
